Color suite: build fix + accessibility + palette adjusters
Build fix
- FluidSimulator.applyConsistency: wrap pow() in Float() to satisfy
  the Float-typed viscosity uniform (was assigning Double to Float)

Color blindness simulation
- ColorTools.simulate(rgb:type:) approximating Brettel/Viénot model
- ColorBlindness enum: protanopia / deuteranopia / tritanopia
- Per-slot 4-up preview in PaletteSlotEditor (Normal + 3 simulations)
- Full-palette preview row in PaletteBuilderView showing all 5 slots
  under each color blindness type

Palette-wide adjustment toolbar
- Cooler / Warmer (rotate hue ±20°)
- Lighten / Darken (brightness ±8%)
- Saturate / Desaturate (saturation ±12%)
- Reverse (flip slot order, locks follow)
- Shuffle (randomize unlocked slot positions)
- All operations respect locked slots

PaletteSlotEditor enhancements
- Color info tiles: RGB / HSB / Hex / Luma (WCAG luminance)
- Color blindness preview chips

Quick actions
- Copy palette as space-separated hex codes to clipboard

+++ PourSimSpike/PourSimSpike/FluidSimulator.swift
@@ -128,7 +128,7 @@ final class FluidSimulator {
     func applyConsistency(_ value: Float) {
         let clamped = max(-1, min(1, value))
         uniforms.damping    = 0.94 - clamped * 0.04   // thin: 0.98 · thick: 0.90
-        uniforms.viscosity  = 0.0008 * pow(3.0, Double(clamped)) // thin: ~0.00027 · thick: ~0.0024
+        uniforms.viscosity  = 0.0008 * Float(pow(3.0, Double(clamped))) // thin: ~0.00027 · thick: ~0.0024
     }
 
     // Fill every grid cell with a solid base color.  Call once before the first frame.
